Add drag and drop function to Each Cell
Plus Adding Available Download List
