+++ src/main/java/ch/hsr/epj/localshare/demo/gui/application/MainWindowController.java
@@ -33,36 +33,6 @@ public class MainWindowController implements Initializable, PeerUpdaterIF {
 	public void initialize(URL location, ResourceBundle resources) {
 		listView.setItems(peerObservableList);
 		listView.setCellFactory(peerListView -> new PeerListViewCell());
-		listView.setOnDragOver(new EventHandler<DragEvent>() {
-
-			@Override
-			public void handle(DragEvent event) {
-				if (event.getGestureSource() != listView
-						&& event.getDragboard().hasFiles()) {
-					/* allow for both copying and moving, whatever user chooses */
-					event.acceptTransferModes(TransferMode.COPY_OR_MOVE);
-				}
-				event.consume();
-			}
-		});
-
-		listView.setOnDragDropped(new EventHandler<DragEvent>() {
-
-			@Override
-			public void handle(DragEvent event) {
-				Dragboard db = event.getDragboard();
-				boolean success = false;
-				if (db.hasFiles()) {
-					System.out.println(db.getFiles().toString());
-					success = true;
-				}
-				/* let the source know whether the string was successfully
-				 * transferred and used */
-				event.setDropCompleted(success);
-
-				event.consume();
-			}
-		});
 
 	}
